+++ CHANGELOG.md
@@ -12,7 +12,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### v9.0.0 require Homebridge v2.0.0
 
 - support for Homebridge v2.0.0
-- add possibility to stop data sampling direct from home app
+- added possibility to disable/enable data refresh direct from home app as a extra tile
+- added possibility to disable/enable data refresh from envoy section
+- added system data refresh sensor
 - code refactor
 - stability and performance improvements
 - dependencies updated